restructure JavaScript plugin code
